use parameterized test in rmsprop
PiperOrigin-RevId: 205914985
